+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 All notable changes to the "yaml-grapher" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## [Unreleased]
 
-- Initial release
+### Added
+- Theme customisation
+
+## [1.0.0] - 2024-11-1
+
+### Added
+
+- Display graph of file references
+- Display conditions on links to file references in the graph
+- Configuration options for graph orientation and theme